+++ src/sdc11073/roles/alarmprovider.py
@@ -32,6 +32,7 @@ class GenericAlarmProvider(providerbase.ProviderRole):
     """
 
     WORKER_THREAD_INTERVAL = 1.0  # seconds
+    self_check_safety_margin = 1.0 # how many seconds before SelfCheckInterval elapses a new self check is performed.
 
     def __init__(self, mdib: ProviderMdib, log_prefix: str):
         super().__init__(mdib, log_prefix)
@@ -409,10 +410,10 @@ class GenericAlarmProvider(providerbase.ProviderRole):
                 alert_system_state = self._mdib.states.descriptor_handle.get_one(alert_system_descr.Handle,
                                                                                  allow_none=True)
                 if alert_system_state is not None:
-                    selfcheck_period = alert_system_descr.SelfCheckPeriod
-                    if selfcheck_period is not None:
-                        last_selfcheck = alert_system_state.LastSelfCheck or 0.0
-                        if time.time() - last_selfcheck >= selfcheck_period:
+                    self_check_period = alert_system_descr.SelfCheckPeriod
+                    if self_check_period is not None:
+                        last_self_check = alert_system_state.LastSelfCheck or 0.0
+                        if time.time() - last_self_check >= self_check_period - self.self_check_safety_margin:
                             states_needing_update.append(alert_system_state)
         except Exception:
             exc = traceback.format_exc()