+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- added a way to proccess operations sync (directly send FINISHED)
+- added a way to process operations directly (directly send 'Fin' instead of Wait, Started,...)
+- added handling of SystemErrorReports.
 
 ### Fixed
 - basic_logging_setup only handles sdc logger, no more side effect due to calling logging.basicConfig. 
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The final OperationInvokedReport has OperationTargetRef parameter set. 
   This required refactoring of Operations handling.
 - moved waveform generator from mdib to roles.waveformprovider
+- alert provider performs self check one second before SelfCheckInterval elapses
 
 ## [2.0.0a6] - 2023-09-11